+++ src/main/java/name/divinityunbound/datagen/ModLootTableProvider.java
@@ -31,6 +31,8 @@ public class ModLootTableProvider extends FabricBlockLootTableProvider {
         addDrop(ModBlocks.SAND_OF_TIME);
         addDrop(ModBlocks.FROZEN_TIME_LAMP);
         addDrop(ModBlocks.DIVINE_TORCH);
+        addDrop(ModBlocks.CELESTIAL_GLASS);
+        addDrop(ModBlocks.DARK_CELESTIAL_GLASS);
 
         addDrop(ModBlocks.CELESTITE_ORE, copperLikeOreDrops(ModBlocks.CELESTITE_ORE, ModItems.RAW_CELESTITE));
         addDrop(ModBlocks.DEEPSLATE_CELESTITE_ORE, copperLikeOreDrops(ModBlocks.DEEPSLATE_CELESTITE_ORE, ModItems.RAW_CELESTITE));