+++ src/main/java/name/divinityunbound/enchantment/BaneOfPassivesEnchantment.java
@@ -1,0 +1,35 @@
+package name.divinityunbound.enchantment;
+
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.enchantment.EnchantmentTarget;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityGroup;
+import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.damage.DamageSource;
+import net.minecraft.entity.damage.DamageSources;
+import net.minecraft.entity.damage.DamageTypes;
+import net.minecraft.entity.passive.PassiveEntity;
+import net.minecraft.server.network.ServerPlayerEntity;
+
+public class BaneOfPassivesEnchantment extends Enchantment {
+    protected BaneOfPassivesEnchantment(Rarity rarity, EnchantmentTarget target, EquipmentSlot... slotTypes) {
+        super(rarity, target, slotTypes);
+    }
+
+    @Override
+    public void onTargetDamaged(LivingEntity user, Entity target, int level) {
+        if (!user.getWorld().isClient()) {
+            if (target instanceof PassiveEntity && user instanceof ServerPlayerEntity) {
+                target.damage(user.getWorld().getDamageSources().playerAttack((ServerPlayerEntity)user), (float)level * 2.5f);
+            }
+        }
+
+        super.onTargetDamaged(user, target, level);
+    }
+
+    @Override
+    public int getMaxLevel() {
+        return 5;
+    }
+}