+++ src/main/java/name/divinityunbound/enchantment/ProsperityEnchantment.java
@@ -1,0 +1,11 @@
+package name.divinityunbound.enchantment;
+
+import net.minecraft.enchantment.EnchantmentTarget;
+import net.minecraft.enchantment.LuckEnchantment;
+import net.minecraft.entity.EquipmentSlot;
+
+public class ProsperityEnchantment extends LuckEnchantment {
+    protected ProsperityEnchantment(Rarity rarity, EnchantmentTarget enchantmentTarget, EquipmentSlot... equipmentSlots) {
+        super(rarity, enchantmentTarget, equipmentSlots);
+    }
+}